+++ tests/SOURCE_GUARDS.md
@@ -14,5 +14,5 @@ extraction.
 | `tests/unit/onboarding/settings.test.ts` | `onboarding-settings-substeps` | Onboarding UI | The multi-substep footer/rendering behavior lives in a Svelte page and requires DOM interaction for true behavior coverage. | Replace with component/browser tests that walk the settings substeps and assert submit button gating. |
 | `tests/unit/plex-login.test.ts` | `plex-login-browser-safe-fields` | Auth client/API | Existing unit tests cover client helpers, but browser-safe response-shape constraints span auth endpoint and redirect load source without a direct exported schema. | Export a response DTO/schema or add route tests that assert completed auth payloads omit Plex tokens/private account fields. |
 | `tests/unit/security/basics.test.ts` | `admin-guard-hook-wiring` | Security/auth hooks | Guard helpers are tested behaviorally; the hook's use of the helper and redirect target are not separately exported. | Add hook-level tests or an exported authorization decision helper covering admin route probes. |
-| `tests/unit/security/dogfood-010-011.test.ts` | `plex-login-no-token-postMessage` | Security/auth client | This is a negative security invariant over absent popup/postMessage logging code; Bun unit tests cannot observe absent future branches without executing browser code. | Replace with browser/client tests that exercise the popup polling path and assert no token payload is posted or logged. |
+| `tests/unit/security/basics.test.ts` | `plex-login-no-token-postMessage` | Security/auth client | This is a negative security invariant over absent popup/postMessage logging code; Bun unit tests cannot observe absent future branches without executing browser code. | Replace with browser/client tests that exercise the popup polling path and assert no token payload is posted or logged. |
 | `tests/unit/sharing/presets.test.ts` | `privacy-preset-not-persisted` | Sharing/privacy UI/actions | Pure preset matching is covered, but the negative persisted/submitted field invariant spans Svelte forms and server action schemas. | Add action/form behavior tests proving preset selection only maps to persisted privacy fields and no `preset` form key is accepted. |
