+++ tests/unit/security/reverse-proxy-diagnostic.test.ts
@@ -1,8 +1,7 @@
 import { afterEach, beforeEach, describe, expect, it } from 'bun:test';
 import { env } from '$env/dynamic/private';
 import { AppSettingsKey, setAppSetting } from '$lib/server/admin/settings.service';
-import { db } from '$lib/server/db/client';
-import { appSettings } from '$lib/server/db/schema';
+import { clearRateLimitStore } from '$lib/server/ratelimit';
 import {
 	assertEnableTrustProxyAllowed,
 	classifySourceAddress,
@@ -11,6 +10,8 @@ import {
 	type ReverseProxyDiagnostic,
 	type ReverseProxyRecommendationAction
 } from '$lib/server/security/reverse-proxy-diagnostic';
+import { GET as diagnosticGET } from '../../../src/routes/api/security/reverse-proxy-diagnostic/+server';
+import { resetSharedTestDb } from '../../helpers/db';
 
 function envRecord(): Record<string, string | undefined> {
 	return env as Record<string, string | undefined>;
@@ -22,7 +23,7 @@ function requestWith(headers: Record<string, string> = {}): Request {
 
 describe('reverse proxy diagnostic', () => {
 	beforeEach(async () => {
-		await db.delete(appSettings);
+		await resetSharedTestDb();
 		delete envRecord().TRUST_PROXY;
 		delete envRecord().ORIGIN;
 	});
@@ -226,15 +227,17 @@ describe('reverse proxy diagnostic', () => {
 		expect(serialized).not.toContain('secret-query');
 	});
 
-	it('classifies source addresses without returning the raw address', () => {
-		expect(classifySourceAddress('127.0.0.1')).toBe('loopback');
-		expect(classifySourceAddress('[::1]')).toBe('loopback');
-		expect(classifySourceAddress('10.0.0.4')).toBe('private-lan');
-		expect(classifySourceAddress('172.20.0.4')).toBe('docker/private-range');
-		expect(classifySourceAddress('100.64.0.1')).toBe('tailscale/cgnat');
-		expect(classifySourceAddress('169.254.10.20')).toBe('link-local');
-		expect(classifySourceAddress('8.8.8.8')).toBe('public');
-		expect(classifySourceAddress('not-an-address')).toBe('unknown');
+	it.each([
+		['127.0.0.1', 'loopback'],
+		['[::1]', 'loopback'],
+		['10.0.0.4', 'private-lan'],
+		['172.20.0.4', 'docker/private-range'],
+		['100.64.0.1', 'tailscale/cgnat'],
+		['169.254.10.20', 'link-local'],
+		['8.8.8.8', 'public'],
+		['not-an-address', 'unknown']
+	] as const)('classifies %s as %s without returning the raw address', (address, category) => {
+		expect(classifySourceAddress(address)).toBe(category);
 	});
 });
 
@@ -274,20 +277,213 @@ describe('assertEnableTrustProxyAllowed', () => {
 		expect(assertEnableTrustProxyAllowed(diagnosticWith('enable'))).toEqual({ ok: true });
 	});
 
-	const blockedActions: ReverseProxyRecommendationAction[] = [
+	it.each([
 		'leave-disabled',
 		'review-proxy',
 		'appears-working',
 		'unable-to-determine',
 		'env-controlled'
-	];
-
-	for (const action of blockedActions) {
-		it(`rejects enabling when the diagnostic recommendation is "${action}"`, () => {
-			expect(assertEnableTrustProxyAllowed(diagnosticWith(action))).toEqual({
-				ok: false,
-				error: ENABLE_TRUST_PROXY_NOT_RECOMMENDED_MESSAGE
-			});
+	] as ReverseProxyRecommendationAction[])('rejects enabling when the diagnostic recommendation is "%s"', (action) => {
+		expect(assertEnableTrustProxyAllowed(diagnosticWith(action))).toEqual({
+			ok: false,
+			error: ENABLE_TRUST_PROXY_NOT_RECOMMENDED_MESSAGE
 		});
-	}
+	});
+});
+
+type HandlerArgs = Parameters<typeof diagnosticGET>[0];
+
+const adminLocals = {
+	user: { id: 1, plexId: 100, username: 'admin', isAdmin: true }
+} as HandlerArgs['locals'];
+
+const userLocals = {
+	user: { id: 2, plexId: 200, username: 'viewer', isAdmin: false }
+} as HandlerArgs['locals'];
+
+function runDiagnosticGET({
+	locals = adminLocals,
+	requestUrl = 'http://internal.local/api/security/reverse-proxy-diagnostic?browserOrigin=https%3A%2F%2Fwrapped.example.com',
+	effectiveUrl,
+	headers = {},
+	ip = '172.18.0.2'
+}: {
+	locals?: HandlerArgs['locals'];
+	requestUrl?: string;
+	effectiveUrl?: string;
+	headers?: Record<string, string>;
+	ip?: string;
+} = {}): ReturnType<typeof diagnosticGET> {
+	const request = new Request(requestUrl, { headers });
+	return diagnosticGET({
+		getClientAddress: () => ip,
+		locals,
+		request,
+		url: new URL(effectiveUrl ?? requestUrl)
+	} as unknown as HandlerArgs);
+}
+
+describe('GET /api/security/reverse-proxy-diagnostic', () => {
+	beforeEach(async () => {
+		clearRateLimitStore();
+		await resetSharedTestDb();
+		delete envRecord().TRUST_PROXY;
+		delete envRecord().ORIGIN;
+	});
+
+	it.each([
+		['anonymous requests with 401', {} as HandlerArgs['locals'], 401, { message: 'Unauthorized' }],
+		['non-admin requests with 403', userLocals, 403, { message: 'Admin access required' }]
+	] as const)('rejects %s', async (_label, locals, status, body) => {
+		const response = await runDiagnosticGET({ locals });
+
+		expect(response.status).toBe(status);
+		expect(response.headers.get('Cache-Control')).toBe('no-store');
+		expect(await response.json()).toEqual(body);
+	});
+
+	it('returns the sanitized diagnostic payload for admins', async () => {
+		const response = await runDiagnosticGET({
+			headers: {
+				'x-forwarded-proto': 'https',
+				'x-forwarded-host': 'wrapped.example.com',
+				'x-forwarded-for': '203.0.113.77',
+				'x-real-ip': '198.51.100.88',
+				forwarded: 'for=secret-client;proto=https;host=hidden.example'
+			}
+		});
+
+		expect(response.status).toBe(200);
+		expect(response.headers.get('Cache-Control')).toBe('no-store');
+		const body = await response.json();
+		expect(Object.keys(body).sort()).toEqual([
+			'forwardedHeaders',
+			'originComparison',
+			'origins',
+			'reasons',
+			'recommendation',
+			'safetyNotice',
+			'sourceAddress',
+			'trustProxy'
+		]);
+		expect(body.recommendation.action).toBe('enable');
+		expect(body.forwardedHeaders).toEqual({
+			present: [
+				'Forwarded',
+				'X-Forwarded-For',
+				'X-Forwarded-Host',
+				'X-Forwarded-Proto',
+				'X-Real-IP'
+			],
+			pair: {
+				status: 'usable',
+				isUsable: true,
+				protoPresent: true,
+				hostPresent: true
+			}
+		});
+		expect(body.sourceAddress).toEqual({ category: 'docker/private-range' });
+	});
+
+	it('compares the browser, raw app, and effective app origins', async () => {
+		await setAppSetting(AppSettingsKey.TRUST_PROXY, 'true');
+
+		const response = await runDiagnosticGET({
+			requestUrl:
+				'http://internal.local/api/security/reverse-proxy-diagnostic?browserOrigin=https%3A%2F%2Fwrapped.example.com',
+			effectiveUrl:
+				'https://wrapped.example.com/api/security/reverse-proxy-diagnostic?browserOrigin=https%3A%2F%2Fwrapped.example.com',
+			headers: {
+				'x-forwarded-proto': 'https',
+				'x-forwarded-host': 'wrapped.example.com'
+			}
+		});
+
+		const body = await response.json();
+		expect(body.origins.rawApp).toBe('http://internal.local');
+		expect(body.origins.effectiveApp).toBe('https://wrapped.example.com');
+		expect(body.origins.browser).toEqual({
+			origin: 'https://wrapped.example.com',
+			isValid: true
+		});
+		expect(body.originComparison).toEqual({
+			browserMatchesRawApp: false,
+			browserMatchesEffectiveApp: true,
+			forwardedPairMatchesBrowser: true
+		});
+		expect(body.recommendation.action).toBe('appears-working');
+	});
+
+	it.each([
+		['missing', 'http://internal.local/api/security/reverse-proxy-diagnostic'],
+		[
+			'invalid',
+			'http://internal.local/api/security/reverse-proxy-diagnostic?browserOrigin=not-a-url'
+		]
+	] as const)('handles %s browserOrigin without failing', async (_label, requestUrl) => {
+		const response = await runDiagnosticGET({
+			requestUrl,
+			headers: {
+				'x-forwarded-proto': 'https',
+				'x-forwarded-host': 'wrapped.example.com'
+			}
+		});
+
+		expect(response.status).toBe(200);
+		const body = await response.json();
+		expect(body.origins.browser).toEqual({ origin: null, isValid: false });
+		expect(body.recommendation.action).toBe('unable-to-determine');
+	});
+
+	it('rejects structurally abusive browserOrigin values', async () => {
+		const response = await runDiagnosticGET({
+			requestUrl: `http://internal.local/api/security/reverse-proxy-diagnostic?browserOrigin=${'a'.repeat(2049)}`
+		});
+
+		expect(response.status).toBe(400);
+		expect(response.headers.get('Cache-Control')).toBe('no-store');
+		expect(await response.json()).toEqual({ message: 'browserOrigin is too long' });
+	});
+
+	it('does not expose raw secrets, forwarded values, source IPs, or query strings', async () => {
+		const response = await runDiagnosticGET({
+			requestUrl:
+				'http://internal.local/api/security/reverse-proxy-diagnostic?browserOrigin=https%3A%2F%2Fbrowser.example.com&token=secret-query',
+			headers: {
+				cookie: 'session=secret-cookie',
+				authorization: 'Bearer secret-authorization',
+				'x-forwarded-proto': 'https',
+				'x-forwarded-host': 'secret-forwarded.example',
+				'x-forwarded-for': '203.0.113.77',
+				'x-real-ip': '198.51.100.88',
+				forwarded: 'for=secret-client;proto=https;host=hidden.example'
+			},
+			ip: '203.0.113.44'
+		});
+
+		const serialized = JSON.stringify(await response.json());
+		for (const secret of [
+			'secret-cookie',
+			'secret-authorization',
+			'secret-forwarded.example',
+			'203.0.113.77',
+			'198.51.100.88',
+			'secret-client',
+			'hidden.example',
+			'203.0.113.44',
+			'secret-query'
+		]) {
+			expect(serialized).not.toContain(secret);
+		}
+	});
+
+	it('rate limits repeated diagnostic requests per admin and source address', async () => {
+		let response: Response | undefined;
+		for (let i = 0; i < 13; i++) response = await runDiagnosticGET();
+
+		expect(response?.status).toBe(429);
+		expect(response?.headers.get('Cache-Control')).toBe('no-store');
+		expect(response?.headers.get('Retry-After')).toBeTruthy();
+		expect(await response?.json()).toEqual({ error: 'Too many diagnostic requests' });
+	});
 });
